move crop code to auto_crop.cc
