+++ CMakeLists.txt
@@ -30,6 +30,7 @@ set(IMGUI_SOURCES
 set(XPANO_SOURCES
   "xpano/main.cc"
   "xpano/algorithm/algorithm.cc"
+  "xpano/algorithm/auto_crop.cc"
   "xpano/algorithm/image.cc"
   "xpano/log/logger.cc"
   "xpano/gui/backends/base.cc"
