+++ xpano/algorithm/algorithm.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <array>
+#include <optional>
 #include <string>
 #include <vector>
 
@@ -81,6 +82,6 @@ StitchResult Stitch(const std::vector<cv::Mat>& images, StitchOptions options);
 
 std::string ToString(cv::Stitcher::Status& status);
 
-utils::RectRRf FindLargestCropRectangle(cv::Mat mask);
+std::optional<utils::RectRRf> FindLargestCrop(cv::Mat mask);
 
 }  // namespace xpano::algorithm
