+++ xpano/algorithm/auto_crop.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <optional>
+
+#include <opencv2/core.hpp>
+
+#include "xpano/utils/rect.h"
+
+namespace xpano::algorithm::crop {
+
+std::optional<utils::RectRRf> FindLargestCrop(cv::Mat mask);
+
+}
